Added to ability to set mtu for lxcbr0
The lxc_host role overrides the configuration
done by the system administrator on the host.

If mtu was defined in /etc/lxc/ and in the
system network interfaces, then these configurations
will be overriden.

This commit should allow the deployer to set the
mtu for its lxcbr0 network, by setting the variable
lxc_net_mtu.

Fixes-Bug: #1518303

Change-Id: I42ab3d0a2c20ae94335de195cdb14579d9f59f5f