Ensure that tasks fail early if there is no ssh public key configured
One of the most common causes of failure in an initial deployment is
that the LXC hosts aren't configured with a ssh public key to
implement into the container caches.

This patch ensures that the play will fatally stop if this
configuration has not been implemented.

Change-Id: I167446eee35e40cde645873fbab7491f840dcd15
(cherry picked from commit aaaaa641b87b896804a101624598b9b55dda53a9)