Correct LXC host public key check
In I167446eee35e40cde645873fbab7491f840dcd15 a pre-flight check
was introduced to validate whether an ssh public key is available
before doing anything on the LXC hosts.

The check is supposed to validate that the requirements for the
lookup are available. The lookup is executed on the deployment
host, not the target host. The check was therefore incorrect.

This patch corrects it to ensure that it does the check in the
right places.

Change-Id: Ifb5b23e2b472bf0c738a01acefba578754f20b4f