Workaround ansible bug of copy delegate_to inventory_hostname
Ansible Bug: https://github.com/ansible/ansible/issues/15929

When ansible 1.9 copy with delegate_to inventory_hostname, the check
for the presence of the source file to copy (the one delegated to
the remote node) is done on the deploy node.
This works fine when the files exist on the deploy node, but will fail
when the deploy node is different than remote node.

This rollbacks the ansible copy task to a command task to fix
the issue we're hitting.

Change-Id: I893e62003bb7bcf8cdc5e6f766b2b67842090a7f
Signed-off-by: Jean-Philippe Evrard <jean-philippe.evrard@rackspace.co.uk>