Update the version of LXC for centos7
This updates the version of LXC to the latest stable (2.0.1).
This is being done because issues exist when running LXC 1.x
within SystemD and how capabilities are handled. To resolve the
capability issues as seen here: [0] the version of LXC has been
updated. This change also pulls Centos 7 in line with the release
series of LXC found on Ubuntu 16.04.

This change will unblock the os_keystone PR [1].

[0] - http://logs.openstack.org/16/320216/10/experimental/gate-openstack-ansible-os_keystone-ansible-func-centos-7-nv/514118f/console.html#_2016-06-01_20_59_54_386
[1] - https://review.openstack.org/#/c/320216

Change-Id: I333fb1887339e8dc9ebf10ff137dda3cff629dc0
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>