Update the version of LXC installed to the latest stable
This change updates the installation of LXC on Ubuntu 14.04 to be the
latest stable release. The change will, at the time of this commit,
use LXC version 2.0.x as the base version. By updating we'll be able
to take advantage of the various performance and stability enhancements
that have been created since the release of 2.x. This also brings our
Ubuntu 14.04 installations in line with Ubuntu 16.04 and CentOS7 such
that we're using a single supported release.

Change-Id: I29220bee11401cc424ef54160cc873b8fc395dd1
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>