Implement primary and secondary keyserver usage for cache prep
Sometimes, especially with restricted internet access through
proxies or firewalls, the LXC container creation fails when using
the download template.

The failure is due to the inability to access the gpg keyserver
and therefore the inability to validate the downloaded image.

This patch implements the usage of a primary and secondary
keyserver, and uses a primary keyserver which is more likely
to be accessible in restricted environments as it's accessed
on port 80.

Change-Id: Ic1ca3d9f3c7d720e9715b6dcc67a7888910e6d0d
Closes-Bug: #1609479