Mask lxc-net.service for systemd
On systemd-based systems, lxc-net is started whenever
lxc is started.  This is because lxc-net is declared
as a dependency of lxc.

We currently mask the service on upstart systems.
We are presently manually configuring the lxcbr0
on all systems, systemd-based systems included.

This patch will mask the lxc-net.service to enable
lxcbr0 to be brought up without error due to conflicting
lxc-net.service.

Closes-Bug: #1580766
Change-Id: I099cf638bbb8874ca45b815383de8588fecbe7b6