Add ability to set upper constraints file
When this role executes the installation of packages, upper
constraints should be used.

This patch adds the ability to set an upper constraint
file URL/path.

The tox config is also adjusted to ensure that it makes
use of upper constraints when installing requirements.

Backport note:
The var lxc_hosts_pip_install_options is not included as it
was added to support CentOS7's LXC build process and is not
necessary to include.

Change-Id: I3cebfe7d8aec84dc533f208f9bdf0a5eb52d9dcd
(cherry picked from commit 3b42216431ed288989e84f0a516e5d3eaccb827c)