Remove unnecessary symlinks from container cache
As far as I can find these symlinks are unnecessary. They
appear to have originally been sourced from the Archlinux
'Linux Containers' wiki page, from where it's now been
removed. There are some noteworthy comments in [1] which
was cited in the wiki page removal.

The bits that are necessary are all included in the LXC
template on a per distribution basis and we now use the
download template in which everything has been prepared.

[1] https://github.com/lxc/lxc/pull/1011

Change-Id: Ia538dc3baea78397b0b13ac4effb6ff1a40fc03c