Allow variable control over lxc_template task enviroment
This allows a deployer to configure a proxy server that is
only applied to these LXC cache tasks. This is sometimes preferable
to configuring such a proxy in the host's enviroment variables. It
may also be preferable to running a private LXC image server and specifying
it using the `lxc_image_cache_server` variable.

Change-Id: I884478c8bf699af2facbc106f745470eaf697ffd