Create SELinux fcontext equivalence for logs
The syslog daemon cannot perform operations on logs within
`/openstack/log` because the files in the directory are labeled as
`default_t`. The syslog daemon does not have a policy that allows
it to touch files with that label.

This patch adds a fcontext equivalence that tells SELinux to consider
`/openstack/log` to be the same as `/var/log`. Proper contexts are
applied to `/openstack/log` as soon as the equivalence is added.

Closes-Bug: 1660429
Change-Id: I7174b5e44f2f84b837aa23507bcf5702d023785e