Use block-rescue for cache create
Instead of using two tasks, skipping the second if the first
succeeds, this will execute the first and totally ignore the
second if the first succeeds. The second will only execute if
the first fails.

This is more efficient to execute and gets rid of the rather
horrible 'failed_when: false' on the first task.

Change-Id: I90f9d95c5dadd0966a04dd7b02b540372d1807cf
(cherry picked from commit 39889dfc2b75c3f039b6bd8da46654063b466095)