Allow variant and extra options to be passed to the cache creation
Currently there is no way to use a variant other than 'default'
without having to override the entire template download options
variable.

This adds a variable to allow the override of the default variant
and another to allow the deployer to add extra options. Examples
include turning off verification, flushing the cache, etc.

This patch provides the option to change it to something
else. This is helpful when pre-building a different default
variant that you wish to use in your own deployments instead
of being forced to rely on the current default.

Change-Id: Id31bdfa34e0ce117d7b64ea9234653ccdf6ac4c6