CentOS: Ensure that GPG keys are imported
In the cache preparation process we copy the host's GPG
keys into the container, but we never import them.

This can result in the following error when trying to
execute the package installation later:

"You have enabled checking of packages via GPG keys. This
is a good thing. However, you do not have any GPG public
keys installed."

Change-Id: I971769e588014ebb814d8549966c9a0c9715155e