Do not use backports repo if it is deliberately not enabled
In https://review.openstack.org/442592 the option to not use
the backports repo for Ubuntu Trusty was added to allow a
deployer to use their own consolidated apt mirror.

This configuration fails due to the default_release argument
in the apt package installation.

This patch ensures that the default_release argument is omitted
if the backports repo addition is deliberately disabled.

Change-Id: If3ec51a4d518bf159f3f13ad5865fb665118b2d1