Cut down default lxc_cache_distro_packages
The current list of packages was the list built up for
Icehouse/Juno/Kilo/Liberty/Mitaka and has never been
cleared out.

As is evidenced by the test results in the integrated
repo [1] this very long list is way over the top and
completely unnecessary.

This patch reduces the list in order to optimise the
build time, reduce maintenance and general bloat.

[1] https://review.openstack.org/446119

Change-Id: I4a797e9cd473da1b7fea03b600fdb86ccac407b3