Copy host environment file into cache
In an environment when a proxy has been configured for internet
access (in /etc/environment), the cache prep will need to also
be configured to make use of it. Currently the onlly way to work
around this is to have a mirror which is accessible without using
a proxy.

This patch ensures that the /etc/environment file is copied into
the LXC cache prior to executing anything in the chroot. This
should ensure that the cache prep can use the same proxy that is
configured on the host.

Change-Id: I4deb7e0a35a0364ceb57b9a12f283ac7439d5e8a