vars: suse-42: Add STARTMODE option for eth0
The default option for STARTMODE is 'manual' which means that the
interface will be setup if ifup is called manually. However, we need
the interface to be activated during boot by wicked so we need to set
this option to 'auto'. Furthermore, the DEVICE option has no effect on
SUSE hosts so we remove it.

Change-Id: Id16e53da96658b21a99588b2fe18a636a96dce7c