Remove EPEL/RDO package installs from cache prep
The keys and yum.repos.d files are copied from the host
and imported prior to the current download/install/config
steps. The distro packages are all installed before this
too, proving that they're already working prior to these
steps.

Obviously these extra commands are not necessary, so they
may as well be removed.

Change-Id: I8ef12ba1f397993376c8d89cfa9a65ebad978b8b