Fix container bridge networking on SUSE
The GATEWAY attribute does not exist on SUSE so we need to add the
default gateway to an ifroute file. Moreover, we drop the DEVICE
attribute which also doesn't exist on SUSE ifcfg files and we add
STARTMODE='auto' to instruct wicked to automatically configure the
interface on every boot. Finally we switch STP to 'off' since there
is only a single bridge accessible from containers.

Change-Id: I9c09e2882614f89cb944a623e7b97298d4bec541