defaults: main: Set default host disk size to 16G
Commit 28c88bb5a84e ("vars: Update container images during bootstrap")
added support for updating the base LXC images during bootstrap.
However, this could now possibly hit the quota limit and lead to
spurious failures depending on the size of the proposed updates.
Commit 18fdecd9d3e1 ("Add option to set the machine image cache size")
added the option to set the disk size for the host but even though the
commit message says that the default option is '16', the one set in the
main.yml file is '8' which is far too low to accommodate the necessary
package updates. We fix this by setting the default value to '16'

Fixes: 28c88bb5a84e ("vars: Update container images during bootstrap")
Fixes: 18fdecd9d3e1 ("Add option to set the machine image cache size")
Change-Id: Id2bce7425e0f44bbe71001e5f2520d00fbc5cf78