tasks: lxc_install_zypper: Add missing files to the dnsmasq apparmor profile
Commit d77bbf6d0ab6 ("tasks: lxc_install_zypper: Relax apparmor
restrictions for dnsmasq") added some neutron directories to the
dnsmasq apparmor profile in order to allow dnsmasq to work on a
neutron environment. However, it missed /var/lib/neutron so apparmor
still prevents dnsmasq from running. Moreover, we fix the regexp
to allow access to files even on neutron subdirectories just to be
sure.

Fixes: d77bbf6d0ab6 ("tasks: lxc_install_zypper: Relax apparmor restrictions for dnsmasq")
Change-Id: I2f839e473267a25c8fd51a0389510485f126fc7c