Add ability to copy more files from LXC host into cache
When preparing the cache, it may be useful to copy more
files into the container cache prior to executing the
package installation. For example, an extra repo config
file may be needed.

This patch allows an optional extra list of files to be
added to the container cache in addition to the existing
set that this defined in vars.

Change-Id: I14a396cda8d1ad23f7ceb61e7867203013f50823