Add dnf support / Use transactions
This patch adds dnf support to the cache prep process.

In addition, this patch changes the yum/dnf cache prep operations
into a single transaction. This is more efficient since it allows
the package manager to solve all of the dependencies in one go and
the package installation happens in one pass (rather than two).

Implements: blueprint centos-and-dnf
Change-Id: I60b6369bf96dfc76918a0cc85f2bb278e2ed2f39