Correct lxc_image_cache_server value
The variable lxc_image_cache_server is shared between
the lxc_container_create role and the lxc_hosts role
in order to make it simpler to override the same value.

The value needs to be a DNS entry, not a URL. This patch
implements a new variable to allow the protocol used to
fetch the images to be set so that the value for
lxc_image_cache_server is uniform between the two roles.

The patch also overrides the value when in nodepool to
the DNS name + path so that the reverse proxy may be
used to improve test success inside nodepool. This is
needed due to poor networking performance when accessing
external sources.

Change-Id: Ia6b149f157e5697d36fdea3dc65cb2ac4815a1a4