Ubuntu: Use dnsmasq-base, not dnsmasq
In Ubuntu the 'dnsmasq' package actually includes
init scripts and service configuration which
conflict with LXC and are best not included. The
actual dependent package is 'dnsmasq-base'.

With this change, the cleanup is no longer
required because the files are removed by the
purge or not installed in the first place.

Closes-Bug: #1711140
Change-Id: I0b0a8dec3943651f379479ea1a3b8038736b9219