Prevent dnsmasq from reading hosts file
OpenStack-Ansible configures /etc/hosts with the management IP
addresses of containers, but dnsmasq gets confused when it sees those
host names associated with a 10.X address on the lxc bridge. It
causes lots of errors in the log files like these:

  not giving name hydrogen_rsyslog_container-307df194 to the DHCP
  lease of 10.0.3.196 because the name exists in /etc/hosts with
  address 172.29.239.71

This patch adds the --no-hosts option for dnsmasq so that it will
stop reading /etc/hosts and filling up log files.

Closes-Bug: 1668949
Change-Id: I7aa2f0081d7d79ab42fcfc28e3ed6839a4f66c8a