tasks: lxc_post_install: Change init.scope pids.max value
Changing the DefaultTasksMax value in /etc/systemd/system.conf
only affects newly started services. However, the init.scope is
there since the beginning so a daemon-reload will not affect it.
As such the value never actually applies to that scope on freshly
installed systems. As a result of which, lets compare the current
value with the desired one and fix it if necessary.

Change-Id: I7548fe4b55ed75846e5fa25f2ce9f3702316f497
(cherry picked from commit a7ea307d46d6ce18ce300c8cb10656fc57d63cc1)