Disable lxc-net on CentOS
The CentOS gate jobs have been plagued by intermittent network
connectivity issues, including failures downloading yum metadata
and failures resolving DNS records. This is happening because
dnsmasq is inaccessible and network paths are broken.

The lxc-net service is trampling on the ifup hooks that run
lxc-system-manage, which takes care of the routes, iptables rules,
and dnsmasq. This patch disables, stops, and masks the lxc-net
service on CentOS to match Ubuntu.

Closes-Bug: 1715248
Change-Id: I4cd10902e367528f59e58b4cb790131d7fbfb7b5