SUSE: Add ability to configure external mirror for openSUSE
Allow deployers to choose a specific mirror by setting the
'lxc_hosts_opensuse_mirror_url' variable

We need to delete all repos which do not match the expected ones so
we need an 'AND' conditional instead of 'OR'. Moreoever, all containers
are expected to have the cloud repository present. This is normally
being installed in the pip_install role but if the host has it already
then preserve that as well.

It's somewhat risky to assume that the host already has the default
repository files named as 'repo-oss' and 'repo-update'. In practise, these
repos could be named anything. If we install the same repo twice because
the filename differs, then strange things might happen. Lets just use
the zypper output to find out if the repos we care about are really
there. We also clean the metadata at the end and we start fresh.

The repositories were missing the .repo extension from the filename so
none of them was copied to the container chroot. Moreover, we should pass
the zypper repository directory as argument to the 'find' command and not
the repository files themselves.

Combined backport of:
- https://review.openstack.org/501650
- https://review.openstack.org/503084
- https://review.openstack.org/507106
- https://review.openstack.org/514850
- https://review.openstack.org/516625

Change-Id: If151501923c5308cfeaee6e33d12df3e7f6959d5