SUSE: Fix repo setup on container cache
We need to delete all repos which do not match the expected ones so
we need an 'AND' conditional instead of 'OR'. Moreoever, all containers
are expected to have the cloud repository present. This is normally
being installed in the pip_install role but if the host has it already
then preserve that as well.

Change-Id: I44bdc79a07afa315e3d037ff05a06f5c79e1f738