Fix inconsistent cache prep timer
The cache prestage step is allowed to take 300 seconds.
A few seconds/minutes (call it X seconds, for the exercice)
go until we test status of the pre-stage.

However, we test 30 times, every 5 seconds (by default), which
makes it a maximum 150 seconds process. So if a download was
started and takes more than 150 + X < 300 (s) it would fail,
where it's expected to succeed, because under 300 seconds.

This enlarges the amount of tests from 30 to 60.
It means that we would check for status from X to 300 + X, which
should be enough to get a failure/success state.

Change-Id: I04e0f05bc079494e6dbb707c6e7cefa983d49667