tasks: lxc_cache_preparation: Log cache preparation command output
We already enable debug output (-x) in the shell script but we never
actually see this output during or after the cache preparation step
takes place. It's useful to log what exactly happened so we can look
for problems or find potential optimization opportunities so just log
the output to /var/log which is also being collected by the various CI
scripts.

Change-Id: I3d0ab31e052889b9ac03de9a3ccba680e840fca5