lxc_cache_preparation: Convert the LXC cache preparation to async task
openSUSE sometimes hangs when running this task. In order to debug this
failure we need to fire up this task and check for it's status on a
separate one. This way, we can avoid Jenkins timeouts and collect logs
properly when that tasks fails to complete within a reasonable amount of
time.

Change-Id: I3acc4436c8b70af097de188b9de2bb9f4e093a09