Remove the continue option for aria2c
The continue option doesn't handle correcly updating file.
When using the --continue option with aria2c if the file exist it will try to
update the current file, if the remote file change (eg.: start the task on
friday and restart it monday when you come back to the office), the remote file
will probably change since image a generated daily and can result in a corrupted
archive.

From the manpage, the continue option is used to resume is to resume a download
which was started by another program. it's not quite the case here.

We could also have a handler to remove files when the task fail but it would
result in the same thing as the file will need to be re-downloaded entirely.

Change-Id: Ia3201e4dba1fa9b99eaff9986fbbdbc37e139b7f
(cherry picked from commit 2a6654398a279dbee3e88f7f58c1ceffbd04b27b)