SUSE: Improve openSUSE repositories handling
It's somewhat risky to assume that the host already has the default
repository files named as 'repo-oss' and 'repo-update'. In practise, these
repos could be named anything. If we install the same repo twice because
the filename differs, then strange things might happen. Lets just use
the zypper output to find out if the repos we care about are really
there. We also clean the metadata at the end and we start fresh.

Change-Id: Ib267ba84d4ff3e4661ec186dcce3cfcc54895f0c