SUSE: Fix copying repositories from the host
The repositories were missing the .repo extension from the filename so
none of them was copied to the container chroot. Moreover, we should pass
the zypper repository directory as argument to the 'find' command and not
the repository files themselves.

To prevent chicken and egg situations, we make the cross-repo test
non-voting.

Change-Id: Iab6f0c408edaf237ad022242c9d6d20257deff1a