SUSE: Avoid repositoy refresh and metadata clean up
The only reason we do a refresh is to import the GPG keys and also to
ensure that our metadata is update. However, we don't really need to do
that since zypper is smart to check for outdated metadata. We can also
import the keys when we install packages or updating the distro.

Change-Id: I4f53f33dd31184d2e3c70c6ad93cea857b30895a