tasks: lxc_cache_create: Use threads when compressing base image
Use multiple threads when compressing the base container image. The
special value of '0' instructs 'xz' to use as many threads as the number
of CPUs on the system. This change improves the total amount of time it
takes to compress the image up to ~85%

Before:
lxc_hosts : Create lxc image ------------------------------------------- 28.42s

After:
lxc_hosts : Create lxc image -------------------------------------------- 3.98s

Change-Id: I5fa162e027882ada4e45a392ae3757f0658cd599