Add package install assertions to SUSE base image
The SUSE base image is using ONLY the leap repositories by default. This
change provides access to the the tumbleweed repos at a far lower
priority giving deployers access to newer packages and features should
they need them without breaking typical LEAP support. One such package
being installed today is "systemd" which gives our container systems
access to systemd 234 which resolves numerous bugs and provides for
some of the mode advanced systemd built-ins we're looking to consume
elsewhere.

In developing this change It was discovered that the flag "--no-confirm"
was not being passed in, which can, in some circumstances cause the
zypper package manager to hang asking for user input. While this flag
is generally assumed given our other options passing it in to the
installation tasks will ensure we're never waiting for user input on
an automated base image build.

Change-Id: Ie214cbd77280321b1795e11020a72dc73cb70c62
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>