Add more testing and increase build timeout for suse/cent
This adds more testing so we're doing everything we can combat
regressions across distros and all of our storage backends. To enure the
greatest possibility of consistent success when running the container prep
process the build timeout has been doubled for both SUSE and Cent.

For the SUSE prep-process the dbus-1 package has been added to the
installation list. This is being done for consistency across the distros
we support.

Depends-On: Ia4856f36b2d106d987e3c774f31493e25a23d4b5
Change-Id: I481cfa36c8f6c5dbda09dadbcb3194c0052ed3f9
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>