Use role defaults for lxc_cache_prep_timeout
In https://review.openstack.org/523525 the default value was
changed for SuSE and CentOS to be higher, but the value was
also overridden using role vars which have a high precedence.

This patch ensures that the value uses the role defaults to
ensure it has the lowest precedence. It also changes the reno
to ensure that the change in default is communicated.

The increased timeout will have no real effect on gating, but
will benefit installations where the mirror is a bit slow.

Change-Id: I41a68313d1841d14001acca591db5c5638e53ffc
(cherry picked from commit b634afe2046e0d4914bac711adbc6b0d9c134b4f)