Allow NetworkManager to work (if present)
This patch adds tasks that check to see if NetworkManager is
installed and running. If it is, the tasks allow NetworkManager
to handle the `lxcbr0` interface.

In addition, the `NetworkManager-wait-online.service` will be
enabled to ensure that all services that depend on networking will
wait for network configuration to be completed.

Partial-Bug: 1738467
Change-Id: I415241daccf22f03826062eea18b3b36b2d9e53e