tasks: lxc_post_install: Handle missing init.scope scope
The init.scope was added in systemd-v226 and support for the pid
controller was added in systemd-v227. However, stable distributions may
be on older systemd version so we need to check whether the scope exists
before we try to adjust the maximum number of processes for that scope.

Change-Id: Ic38e17d110e7a72770286c2faf571ba03bde7f32
Link: https://github.com/systemd/systemd/commit/efdb02375beb0a940c3320865572913780b4d7de
Link: https://github.com/systemd/systemd/commit/03a7b521e3ffb7f5d153d90480ba5d4bc29d1e8f
Closes-Bug: #1743382
(cherry picked from commit 165bd142b1d40467284f6b13457888b599b2bebd)